+++ packages/polaris-viz/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Added
+
+- Set `enableHideOverflow` on `<LegendContainer />` in `<HorizontalBarChart />` to render `+ X more` when legends flow outside the chart bounds.
 
 ## [12.0.0] - 2024-03-25